satisfy marco
by putting the thing into one file
